Move UserSpaceMapping interface to types folder and move SpaceWithAccessTokenResponse interface to SpaceClient.ts
