+++ ui/src/AccountDropdown/ShareAccessForm/InviteEditorsFormSection/InviteEditorsFormSection.tsx
@@ -19,7 +19,7 @@ import React, {CSSProperties, FormEvent, useEffect, useState} from 'react';
 import SpaceClient from 'Space/SpaceClient';
 import FormButton from 'ModalFormComponents/FormButton';
 import {Space} from 'Types/Space';
-import {UserSpaceMapping} from 'Space/UserSpaceMapping';
+import {UserSpaceMapping} from 'Types/UserSpaceMapping';
 
 import UserAccessList from 'ReusableComponents/UserAccessList';
 import Creatable from 'react-select/creatable';
