+++ ui/src/ReusableComponents/UserAccessList.tsx
@@ -23,7 +23,7 @@ import {
 } from '../ModalFormComponents/ReactSelectStyles';
 import React, {CSSProperties, useState} from 'react';
 import {Space} from 'Types/Space';
-import {UserSpaceMapping} from '../Space/UserSpaceMapping';
+import {UserSpaceMapping} from '../Types/UserSpaceMapping';
 
 import './UserAccessList.scss';
 import SpaceClient from '../Space/SpaceClient';
