+++ ui/src/SpaceDashboard/SpaceDashboardTile.test.tsx
@@ -21,7 +21,7 @@ import TestData from '../Utils/TestData';
 import React from 'react';
 import SpaceDashboardTile from './SpaceDashboardTile';
 import SpaceClient from '../Space/SpaceClient';
-import {UserSpaceMapping} from '../Space/UserSpaceMapping';
+import {UserSpaceMapping} from '../Types/UserSpaceMapping';
 import {CurrentUserState} from '../State/CurrentUserState';
 import {ModalContents, ModalContentsState} from '../State/ModalContentsState';
 import {RecoilObserver} from '../Utils/RecoilObserver';
